fix: address review findings #005, #009, #011 — extract hooks from NeighborhoodPage
Break NeighborhoodPage (18 useState, 445 lines) into four focused hooks:
- useNeighborhoodData: map data fetching (libraries, rec centers, transit, boundaries)
- useCommunityMetrics: 311 metrics, demographics, transit score, access gap
- useReportGeneration: report auto-fetch, on-demand generation, state
- useBlockData: block-level 311 data for pinned locations

Also fixes #009 (report effect now gates on non-null metrics, eliminating
redundant getPreGeneratedReport calls) and #011 (all fetch errors now
surface to user via dataError state instead of silent console.error).
