+++ src/pages/neighborhood-page.tsx
@@ -1,44 +1,19 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { useParams, useNavigate, Link } from 'react-router-dom';
 import SanDiegoMap from '../components/map/san-diego-map';
 import NeighborhoodSelector from '../components/ui/neighborhood-selector';
 import Sidebar from '../components/ui/sidebar';
 import { FlyerLayout } from '../components/flyer/flyer-layout';
 import { PrintFlyerFab } from '../components/flyer/print-flyer-fab';
-import { getLibraries, getRecCenters, getTransitStops, get311, getDemographics, generateReport, getPreGeneratedReport, getNeighborhoodBoundaries, getTransitScore, getAccessGap, getBlockData } from '../api/client';
-import type { BlockMetrics, CommunityAnchor, CommunityReport, NeighborhoodProfile, TransitStop } from '../types';
-import type { FeatureCollection } from 'geojson';
+import type { CommunityAnchor } from '../types';
 import { useLanguage } from '../i18n/context';
 import { SUPPORTED_LANGUAGES } from '../i18n/translations';
 import { toSlug, fromSlug } from '../utils/slug';
 import { findCommunityAtPoint } from '../utils/point-in-polygon';
-
-const DEFAULT_TRANSIT: NeighborhoodProfile['transit'] = {
-  nearbyStopCount: 0, nearestStopDistance: 0, stopCount: 0, agencyCount: 0,
-  agencies: [], transitScore: 0, cityAverage: 0, travelTimeToCityHall: null,
-};
-
-function buildDefaultAnchor(community: string): CommunityAnchor {
-  return { id: '', name: community, type: 'library' as const, lat: 0, lng: 0, address: '', community };
-}
-
-function buildProfile(
-  community: string,
-  anchor: CommunityAnchor | null,
-  metrics: NeighborhoodProfile['metrics'],
-  transitScore: NeighborhoodProfile['transit'] | null,
-  topLanguages: { language: string; percentage: number }[],
-  accessGap: NeighborhoodProfile['accessGap'],
-): NeighborhoodProfile {
-  return {
-    communityName: community,
-    anchor: anchor ?? buildDefaultAnchor(community),
-    metrics,
-    transit: transitScore ?? DEFAULT_TRANSIT,
-    demographics: { topLanguages },
-    accessGap: accessGap ?? null,
-  };
-}
+import { useNeighborhoodData } from '../hooks/use-neighborhood-data';
+import { useCommunityMetrics } from '../hooks/use-community-metrics';
+import { useReportGeneration } from '../hooks/use-report-generation';
+import { useBlockData } from '../hooks/use-block-data';
 
 export default function NeighborhoodPage() {
   const { slug } = useParams<{ slug: string }>();
@@ -46,31 +21,26 @@ export default function NeighborhoodPage() {
   const { lang, setLang, t, reportLang } = useLanguage();
   const [mobileView, setMobileView] = useState<'map' | 'info'>('map');
 
-  const [libraries, setLibraries] = useState<CommunityAnchor[]>([]);
-  const [recCenters, setRecCenters] = useState<CommunityAnchor[]>([]);
-  const [transitStops, setTransitStops] = useState<TransitStop[]>([]);
-  const [neighborhoodBoundaries, setNeighborhoodBoundaries] = useState<FeatureCollection | null>(null);
-
   const [selectedCommunity, setSelectedCommunity] = useState<string | null>(
     slug ? fromSlug(slug) : null,
   );
   const [selectedAnchor, setSelectedAnchor] = useState<CommunityAnchor | null>(null);
-  const [metrics, setMetrics] = useState<NeighborhoodProfile['metrics'] | null>(null);
-  const [metricsLoading, setMetricsLoading] = useState(false);
-  const [topLanguages, setTopLanguages] = useState<{ language: string; percentage: number }[]>([]);
-
-  const [transitScore, setTransitScore] = useState<NeighborhoodProfile['transit'] | null>(null);
-  const [accessGap, setAccessGap] = useState<NeighborhoodProfile['accessGap']>(null);
   const [pinnedLocation, setPinnedLocation] = useState<{ lat: number; lng: number } | null>(null);
-  const [blockData, setBlockData] = useState<BlockMetrics | null>(null);
-  const [blockLoading, setBlockLoading] = useState(false);
   const [blockRadius, setBlockRadius] = useState(0.25);
 
-  const [dataError, setDataError] = useState<string | null>(null);
-
-  const [report, setReport] = useState<CommunityReport | null>(null);
-  const [reportLoading, setReportLoading] = useState(false);
-  const [reportError, setReportError] = useState<string | null>(null);
+  // Custom hooks — each manages its own state group
+  const { libraries, recCenters, transitStops, neighborhoodBoundaries, dataError } = useNeighborhoodData();
+  const { metrics, metricsLoading, topLanguages, transitScore, accessGap } = useCommunityMetrics(selectedCommunity);
+  const { report, reportLoading, reportError, handleGenerateReport } = useReportGeneration({
+    selectedCommunity,
+    selectedAnchor,
+    metrics,
+    transitScore,
+    topLanguages,
+    accessGap,
+    reportLang,
+  });
+  const { blockData, setBlockData, blockLoading } = useBlockData(pinnedLocation, blockRadius);
 
   // Sync URL -> state when slug changes (e.g. browser back/forward)
   useEffect(() => {
@@ -81,129 +51,6 @@ export default function NeighborhoodPage() {
     }
   }, [slug]);
 
-  // Fetch map data on mount
-  useEffect(() => {
-    getLibraries().then(setLibraries).catch((err) => { console.error(err); setDataError('Failed to load map data'); });
-    getRecCenters().then(setRecCenters).catch(console.error);
-    getNeighborhoodBoundaries().then(setNeighborhoodBoundaries).catch(console.error);
-    getTransitStops()
-      .then(setTransitStops)
-      .catch(console.error);
-  }, []);
-
-  // Fetch 311 metrics and demographics when community changes
-  useEffect(() => {
-    if (!selectedCommunity) {
-      setMetrics(null);
-      setReport(null);
-      setTopLanguages([]);
-      setTransitScore(null);
-      setAccessGap(null);
-      return;
-    }
-
-    const controller = new AbortController();
-    const { signal } = controller;
-
-    setMetricsLoading(true);
-    setMetrics(null);
-    setTopLanguages([]);
-    setTransitScore(null);
-    setAccessGap(null);
-
-    get311(selectedCommunity)
-      .then((data) => { if (!signal.aborted) setMetrics(data); })
-      .catch((err) => { if (!signal.aborted) console.error(err); })
-      .finally(() => { if (!signal.aborted) setMetricsLoading(false); });
-
-    getTransitScore(selectedCommunity)
-      .then((data) => { if (!signal.aborted) setTransitScore(data); })
-      .catch(() => { /* transit score may not be available */ });
-
-    getAccessGap(selectedCommunity)
-      .then((data) => { if (!signal.aborted && data?.accessGapScore != null) setAccessGap(data); })
-      .catch(() => { /* access gap score may not be available */ });
-
-    // Try to fetch demographics for language suggestion
-    getDemographics(selectedCommunity)
-      .then((data) => {
-        if (!signal.aborted && data?.topLanguages) setTopLanguages(data.topLanguages);
-      })
-      .catch(() => {
-        // Demographics may not be available for all communities
-      });
-
-    return () => { controller.abort(); };
-  }, [selectedCommunity]);
-
-  // Clear report when community or language changes
-  const generatingRef = useRef(false);
-  useEffect(() => {
-    setReport(null);
-    setReportError(null);
-  }, [selectedCommunity, reportLang]);
-
-  // Track whether we already have a report to avoid stale closure reads
-  const hasReportRef = useRef(false);
-  useEffect(() => { hasReportRef.current = report !== null; }, [report]);
-
-  // Auto-fetch pre-generated report, falling back to on-demand generation
-  useEffect(() => {
-    if (!selectedCommunity) return;
-    if (generatingRef.current) return;
-
-    let cancelled = false;
-    setReportLoading(true);
-
-    (async () => {
-      // Step 1: Try to load pre-generated report instantly
-      const cached = await getPreGeneratedReport(selectedCommunity, reportLang);
-      if (cancelled) return;
-
-      if (cached) {
-        setReport(cached);
-        setReportLoading(false);
-        return;
-      }
-
-      // Step 2: No cached report — generate on-demand if metrics are available
-      if (!metrics) {
-        // Metrics haven't loaded yet; this effect will re-run when they do
-        setReportLoading(false);
-        return;
-      }
-
-      // Already have a report — don't regenerate on metrics updates
-      if (hasReportRef.current) {
-        setReportLoading(false);
-        return;
-      }
-
-      generatingRef.current = true;
-
-      const profile = buildProfile(selectedCommunity, selectedAnchor, metrics, transitScore, topLanguages, accessGap);
-
-      try {
-        const result = await generateReport(profile, reportLang);
-        if (!cancelled) setReport(result);
-      } catch (err) {
-        if (!cancelled) setReportError(err instanceof Error ? err.message : 'Failed to generate report');
-      } finally {
-        generatingRef.current = false;
-        if (!cancelled) setReportLoading(false);
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-      generatingRef.current = false;
-    };
-  // Intentionally omitting selectedAnchor, transitScore, topLanguages, accessGap —
-  // report generation should only trigger on community/language/metrics changes.
-  // Secondary data is captured at call time but should not re-trigger generation.
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedCommunity, reportLang, metrics]);
-
   const handleCommunityChange = useCallback(
     (community: string) => {
       if (community) {
@@ -213,11 +60,10 @@ export default function NeighborhoodPage() {
       }
       setSelectedCommunity(community || null);
       setSelectedAnchor(null);
-      // Clear block data when user manually changes neighborhood via dropdown
       setPinnedLocation(null);
       setBlockData(null);
     },
-    [navigate],
+    [navigate, setBlockData],
   );
 
   const handleAnchorClick = useCallback(
@@ -233,7 +79,6 @@ export default function NeighborhoodPage() {
     setPinnedLocation({ lat, lng });
     setBlockData(null);
 
-    // Auto-detect enclosing neighborhood from GeoJSON boundaries
     if (neighborhoodBoundaries) {
       const detected = findCommunityAtPoint(lat, lng, neighborhoodBoundaries);
       if (detected && detected !== selectedCommunity) {
@@ -241,37 +86,7 @@ export default function NeighborhoodPage() {
         navigate(`/neighborhood/${toSlug(detected)}`);
       }
     }
-  }, [neighborhoodBoundaries, selectedCommunity, navigate]);
-
-  // Re-fetch block data when radius changes
-  useEffect(() => {
-    if (!pinnedLocation) return;
-    let cancelled = false;
-    setBlockLoading(true);
-    getBlockData(pinnedLocation.lat, pinnedLocation.lng, blockRadius)
-      .then((data) => { if (!cancelled) setBlockData(data); })
-      .catch((err) => { if (!cancelled) console.error('Failed to fetch block data', err); })
-      .finally(() => { if (!cancelled) setBlockLoading(false); });
-    return () => { cancelled = true; };
-  }, [blockRadius, pinnedLocation]);
-
-  const handleGenerateReport = useCallback(async (language: string) => {
-    if (!selectedCommunity || !metrics) return;
-
-    const profile = buildProfile(selectedCommunity, selectedAnchor, metrics, transitScore, topLanguages, accessGap);
-
-    setReportLoading(true);
-    setReportError(null);
-    try {
-      const result = await generateReport(profile, language);
-      setReport(result);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'Failed to generate report';
-      setReportError(message);
-    } finally {
-      setReportLoading(false);
-    }
-  }, [selectedCommunity, selectedAnchor, metrics, topLanguages, transitScore, accessGap]);
+  }, [neighborhoodBoundaries, selectedCommunity, navigate, setBlockData]);
 
   return (
     <div className="flex flex-col h-full md:flex-row print:block">
